Added user registration, login, email verification, home page, bug fixes and other code improvements

+++ src/Controller/BooksController.php
@@ -17,15 +17,14 @@ use Symfony\Component\EventDispatcher\EventDispatcher;
 use App\Event\BookAddEvent;
 use App\EventListener\BookAddEventListener;
 
-#[Route('/books')]
+#[Route('/books', name: 'app_books_')]
 final class BooksController extends AbstractController
 {
-  #[Route(name: 'app_books_index', methods: ['GET'])]
+  #[Route(name: 'index', methods: ['GET'])]
   public function index(
     Request $request,
     BooksRepository $booksRepository
-  ): Response
-  {
+  ): Response {
     $currentPage = (int) ($request->get('page') ?? 1);
     $searchValue = $request->get('book-title-or-author');
 
@@ -45,32 +44,33 @@ final class BooksController extends AbstractController
     ]);
   }
 
-  #[Route('/{id}', name: 'app_books_show', methods: ['GET'])]
+  #[Route('/{id}', name: 'show', methods: ['GET'], requirements: ['id' => '\d+'])]
   public function show(Book $book): Response
   {
     return $this->render('books/show.html.twig', [
-        'book' => $book,
+      'book' => $book,
     ]);
   }
 
-  #[Route('/new', name: 'app_books_new', methods: ['GET', 'POST'])]
+  #[Route('/new', name: 'new', methods: ['GET', 'POST'], priority: 2)]
   public function new(
     Request $request,
     EntityManagerInterface $entityManager,
     BookAddEventListener $listener
-  ): Response
-  {
+  ): Response {
+    $this->denyAccessUnlessGranted('IS_AUTHENTICATED');
+
     $book = new Book();
     $form = $this->createForm(BookType::class, $book);
     $form->handleRequest($request);
 
     if ($form->isSubmitted() && $form->isValid()) {
       $entityManager->persist($book);
+      $bookOriginal = null;
 
-      $result = $this->store($form, $book, null, $entityManager);
+      $result = $this->store($form, $book, $bookOriginal, $entityManager);
 
       if ($result) {
-        // $listener = $this->container->get(BookAddEventListener::class);
         $dispatcher = new EventDispatcher();
 
         $dispatcher->addListener(BookAddEvent::NAME, [$listener, 'onBookAdd']);
@@ -86,17 +86,18 @@ final class BooksController extends AbstractController
 
     return $this->render('books/new.html.twig', [
       'book' => $book,
-      'form' => $form,
+      'form' => $form
     ]);
   }
 
-  #[Route('/{id}/edit', name: 'app_books_edit', methods: ['GET', 'POST'])]
+  #[Route('/{id}/edit', name: 'edit', methods: ['GET', 'POST'])]
   public function edit(
     Request $request,
     Book $book,
     EntityManagerInterface $entityManager
-  ): Response
-  {
+  ): Response {
+    $this->denyAccessUnlessGranted('IS_AUTHENTICATED');
+
     $bookOriginal = clone $book;
     $form = $this->createForm(BookType::class, $book);
     $form->handleRequest($request);
@@ -113,19 +114,20 @@ final class BooksController extends AbstractController
 
     return $this->render('books/edit.html.twig', [
       'book' => $bookOriginal,
-      'form' => $form,
+      'form' => $form
     ]);
   }
 
-  #[Route('/{id}', name: 'app_books_delete', methods: ['POST'])]
+  #[Route('/{id}', name: 'delete', methods: ['POST'])]
   public function delete(
     Request $request,
     Book $book,
     EntityManagerInterface $entityManager
-  ): Response
-  {
+  ): Response {
+    $this->denyAccessUnlessGranted('IS_AUTHENTICATED');
+
     if ($this->isCsrfTokenValid(
-      'delete'.$book->getId(),
+      'delete' . $book->getId(),
       $request->getPayload()->getString('_token')
     )) {
       $entityManager->remove($book);
@@ -149,9 +151,9 @@ final class BooksController extends AbstractController
     $deletePhoto = $form->has('deletePhoto') ? $form->get('deletePhoto')->getData() : false;
     $photo = $bookOriginal?->getPhoto();
 
-    $book->setPhoto(match(true) {
+    $book->setPhoto(match (true) {
       $deletePhoto => null,
-      $newPhoto != null => 'book'.'-'.bin2hex(random_bytes(13)).'.'.$newPhoto->guessExtension(),
+      $newPhoto != null => 'book' . '-' . bin2hex(random_bytes(13)) . '.' . $newPhoto->guessExtension(),
       default => $photo
     });
 
@@ -164,10 +166,7 @@ final class BooksController extends AbstractController
     $photoDir = $book->getSystemPhotosDir();
 
     if ($photo && ($deletePhoto || $newPhoto)) {
-      $photo = str_replace(['/', '..', '\\'], '', $photo);
-      $photoPath = str_replace('..', '', $photoDir)."/$photo";
-
-      $this->deletePhoto($photoPath);
+      $this->deletePhoto("$photoDir/$photo");
     }
 
     if ($newPhoto) {
@@ -181,7 +180,10 @@ final class BooksController extends AbstractController
     return true;
   }
 
-  private function deletePhoto(string $photoFullPath): bool {
+  private function deletePhoto(string $photoFullPath): bool
+  {
+    $photoFullPath = str_replace(['..', '\\'], '', $photoFullPath);
+
     if (file_exists($photoFullPath) && is_file($photoFullPath)) {
       return unlink($photoFullPath);
     }
